[STYLE] rename user file
