+++ functions/api/routes/user/userMajorListGET.js
