+++ functions/api/routes/user/userUniversityListGET.js
